# backend/src/main/java/it/unical/ea/Travel/DTOs/favorite/FavoriteListResponseDTO.java
package it.unical.ea.Travel.DTOs.favorite;

import java.time.LocalDateTime;
import java.util.UUID;

import it.unical.ea.Travel.Entities.favorite.FavoriteList.Visibility;

public record FavoriteListResponseDTO(
        UUID id,
        UUID ownerId,
        String name,
        String description,
        Visibility visibility,
        LocalDateTime createdAt,
        LocalDateTime updatedAt) {
}