# frontend/app/specs/openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/user":{"get":{"tags":["user-controller"],"operationId":"getUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}}},"post":{"tags":["user-controller"],"operationId":"saveUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}},"delete":{"tags":["user-controller"],"operationId":"deleteUser","parameters":[{"name":"stringId","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/location":{"get":{"tags":["location-controller"],"operationId":"getLocations","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LocationResponseDTO"}}}}}}},"post":{"tags":["location-controller"],"operationId":"saveLocation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LocationRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LocationResponseDTO"}}}}}}},"/experience":{"get":{"tags":["experience-controller"],"operationId":"getExperiences","parameters":[{"name":"organizerId","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ExperienceResponseDTO"}}}}}}},"post":{"tags":["experience-controller"],"operationId":"saveExperience","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExperienceRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExperienceResponseDTO"}}}}}}},"/experience-stop":{"get":{"tags":["experience-stop-controller"],"operationId":"getExperienceStops","parameters":[{"name":"experienceId","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ExperienceStopResponseDTO"}}}}}}},"post":{"tags":["experience-stop-controller"],"operationId":"saveExperienceStop","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExperienceStopRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExperienceStopResponseDTO"}}}}}}},"/experience-image":{"get":{"tags":["experience-image-controller"],"operationId":"getExperienceImages","parameters":[{"name":"experienceId","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ExperienceImageResponseDTO"}}}}}}},"post":{"tags":["experience-image-controller"],"operationId":"saveExperienceImage","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExperienceImageRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExperienceImageResponseDTO"}}}}}}},"/api/auth/signup":{"post":{"tags":["auth-controller"],"operationId":"signup","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/FavoriteList":{"get":{"tags":["favorite-list-controller"],"operationId":"getFavoriteLists","parameters":[{"name":"ownerId","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FavoriteListResponseDTO"}}}}}}},"post":{"tags":["favorite-list-controller"],"operationId":"saveFavoriteList","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FavoriteListRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FavoriteListResponseDTO"}}}}}},"delete":{"tags":["favorite-list-controller"],"operationId":"deleteFavoriteList","parameters":[{"name":"stringIdString","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/FavoriteListItem":{"get":{"tags":["favorite-list-item-controller"],"operationId":"getFavoriteListItems","parameters":[{"name":"favoriteListId","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FavoriteListItemResponseDTO"}}}}}}},"post":{"tags":["favorite-list-item-controller"],"operationId":"saveFavoriteListItem","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FavoriteListItemRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FavoriteListItemResponseDTO"}}}}}}},"/user/{stringId}":{"get":{"tags":["user-controller"],"operationId":"getUser","parameters":[{"name":"stringId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/location/{stringId}":{"get":{"tags":["location-controller"],"operationId":"getLocation","parameters":[{"name":"stringId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LocationResponseDTO"}}}}}},"delete":{"tags":["location-controller"],"operationId":"deleteLocation","parameters":[{"name":"stringId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/experience/{stringId}":{"get":{"tags":["experience-controller"],"operationId":"getExperience","parameters":[{"name":"stringId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExperienceResponseDTO"}}}}}},"delete":{"tags":["experience-controller"],"operationId":"deleteExperience","parameters":[{"name":"stringId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/experience-stop/{stringId}":{"get":{"tags":["experience-stop-controller"],"operationId":"getExperienceStop","parameters":[{"name":"stringId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExperienceStopResponseDTO"}}}}}},"delete":{"tags":["experience-stop-controller"],"operationId":"deleteExperienceStop","parameters":[{"name":"stringId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/experience-image/{stringId}":{"get":{"tags":["experience-image-controller"],"operationId":"getExperienceImage","parameters":[{"name":"stringId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ExperienceImageResponseDTO"}}}}}},"delete":{"tags":["experience-image-controller"],"operationId":"deleteExperienceImage","parameters":[{"name":"stringId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/api/basic":{"get":{"tags":["keycloak-test-controller"],"operationId":"basic","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/admin":{"get":{"tags":["keycloak-test-controller"],"operationId":"admin","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/FavoriteListItem/{stringId}":{"get":{"tags":["favorite-list-item-controller"],"operationId":"getFavoriteListItem","parameters":[{"name":"stringId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FavoriteListItemResponseDTO"}}}}}},"delete":{"tags":["favorite-list-item-controller"],"operationId":"deleteFavoriteListItem","parameters":[{"name":"stringId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/FavoriteList/{stringId}":{"get":{"tags":["favorite-list-controller"],"operationId":"getFavoriteList","parameters":[{"name":"stringId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FavoriteListResponseDTO"}}}}}}}},"components":{"schemas":{"SignupRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1},"password":{"type":"string","maxLength":2147483647,"minLength":8,"pattern":"^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d).*$"},"firstName":{"type":"string","maxLength":100,"minLength":0},"lastName":{"type":"string","maxLength":100,"minLength":0}},"required":["email","firstName","lastName","password"]},"User":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"email":{"type":"string"},"passwordHash":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"avatarUrl":{"type":"string"},"emailVerified":{"type":"boolean"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"deletedAt":{"type":"string","format":"date-time"},"roles":{"type":"string"},"deleted":{"type":"boolean"}}},"LocationRequestDTO":{"type":"object","properties":{"name":{"type":"string","maxLength":255,"minLength":0},"address":{"type":"string","maxLength":255,"minLength":0},"city":{"type":"string","maxLength":100,"minLength":0},"country":{"type":"string","maxLength":100,"minLength":0},"latitude":{"type":"number","maximum":90.0,"minimum":-90.0},"longitude":{"type":"number","maximum":180.0,"minimum":-180.0}},"required":["city","country","name"]},"LocationResponseDTO":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"address":{"type":"string"},"city":{"type":"string"},"country":{"type":"string"},"latitude":{"type":"number"},"longitude":{"type":"number"}}},"ExperienceRequestDTO":{"type":"object","properties":{"organizerId":{"type":"string","format":"uuid"},"type":{"type":"string","enum":["TRIP","ACTIVITY","CUSTOM_ITINERARY"]},"title":{"type":"string","maxLength":255,"minLength":0},"description":{"type":"string","maxLength":5000,"minLength":0},"basePrice":{"type":"number"},"currency":{"type":"string","minLength":1,"pattern":"^[A-Z]{3}$"},"maxParticipants":{"type":"integer","format":"int32"},"minParticipants":{"type":"integer","format":"int32"},"durationMinutes":{"type":"integer","format":"int32"},"startDate":{"type":"string","format":"date"},"endDate":{"type":"string","format":"date"},"coverImageUrl":{"type":"string","maxLength":500,"minLength":0},"status":{"type":"string","enum":["DRAFT","PUBLISHED","CANCELLED","ARCHIVED"]}},"required":["basePrice","currency","organizerId","status","title","type"]},"ExperienceOrganizerDTO":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"firstName":{"type":"string"},"lastName":{"type":"string"},"avatarUrl":{"type":"string"}}},"ExperienceResponseDTO":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"organizer":{"$ref":"#/components/schemas/ExperienceOrganizerDTO"},"type":{"type":"string","enum":["TRIP","ACTIVITY","CUSTOM_ITINERARY"]},"title":{"type":"string"},"description":{"type":"string"},"basePrice":{"type":"number"},"currency":{"type":"string"},"maxParticipants":{"type":"integer","format":"int32"},"minParticipants":{"type":"integer","format":"int32"},"durationMinutes":{"type":"integer","format":"int32"},"startDate":{"type":"string","format":"date"},"endDate":{"type":"string","format":"date"},"coverImageUrl":{"type":"string"},"status":{"type":"string","enum":["DRAFT","PUBLISHED","CANCELLED","ARCHIVED"]},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"ExperienceStopRequestDTO":{"type":"object","properties":{"experienceId":{"type":"string","format":"uuid"},"sequenceOrder":{"type":"integer","format":"int32"},"title":{"type":"string","maxLength":255,"minLength":0},"description":{"type":"string","maxLength":5000,"minLength":0},"locationId":{"type":"string","format":"uuid"},"arrivalTime":{"type":"string"},"departureTime":{"type":"string"},"durationMinutes":{"type":"integer","format":"int32"}},"required":["experienceId","locationId","sequenceOrder","title"]},"ExperienceStopResponseDTO":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"experienceId":{"type":"string","format":"uuid"},"sequenceOrder":{"type":"integer","format":"int32"},"title":{"type":"string"},"description":{"type":"string"},"locationId":{"type":"string","format":"uuid"},"locationName":{"type":"string"},"arrivalTime":{"type":"string"},"departureTime":{"type":"string"},"durationMinutes":{"type":"integer","format":"int32"}}},"ExperienceImageRequestDTO":{"type":"object","properties":{"experienceId":{"type":"string","format":"uuid"},"url":{"type":"string","maxLength":500,"minLength":0},"displayOrder":{"type":"integer","format":"int32"}},"required":["displayOrder","experienceId","url"]},"ExperienceImageResponseDTO":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"experienceId":{"type":"string","format":"uuid"},"url":{"type":"string"},"displayOrder":{"type":"integer","format":"int32"}}},"LoginRequest":{"type":"object","properties":{"email":{"type":"string","format":"email"},"password":{"type":"string","maxLength":2147483647,"minLength":8,"pattern":"^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d).*$"}},"required":["password"]},"FavoriteListRequestDTO":{"type":"object","properties":{"ownerId":{"type":"string","format":"uuid"},"name":{"type":"string","minLength":1},"description":{"type":"string","maxLength":1000,"minLength":0},"visibility":{"type":"string","enum":["PRIVATE","SHARED"]}},"required":["name","ownerId","visibility"]},"FavoriteListResponseDTO":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"ownerId":{"type":"string","format":"uuid"},"name":{"type":"string"},"description":{"type":"string"},"visibility":{"type":"string","enum":["PRIVATE","SHARED"]},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"FavoriteListItemRequestDTO":{"type":"object","properties":{"favoriteListId":{"type":"string","format":"uuid"},"experienceId":{"type":"string","format":"uuid"},"notes":{"type":"string","maxLength":1000,"minLength":0}},"required":["experienceId","favoriteListId"]},"FavoriteListItemResponseDTO":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"favoriteList":{"$ref":"#/components/schemas/FavoriteListResponseDTO"},"experience":{"$ref":"#/components/schemas/ExperienceResponseDTO"},"addedAt":{"type":"string","format":"date-time"},"notes":{"type":"string"}}}}}}